Move the skill content processing to a separate file + add a test
