+++ server/fixtures/example-skill.md
@@ -1,0 +1,11 @@
+---
+name: test-skill
+description: This is an example of a SKILL.md file for testing purposes.
+---
+# Testing Skill
+
+This skill helps you perform tests on the parseSkillMarkdown function.
+
+## Second heading
+
+This information should not be included in the description.
